# src/main/webapp/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "RevolutTransferAccount APIs",
    "version" : "1.0.0",
    "title" : "RevolutTransferAccount",
    "contact" : {
      "name" : "Kalmanbetov Kanat",
      "url" : "https://github.com/kabakan/TransferAccount/",
      "email" : "kanat.kaba@gmail.com"
    }
  },
  "host" : "localhost:8587",
  "basePath" : "/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/api/accounts/delete" : {
      "get" : {
        "tags" : [ "Accounts" ],
        "summary" : "Delete accounts",
        "description" : "",
        "operationId" : "delete",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/api/accounts/get" : {
      "get" : {
        "tags" : [ "Accounts" ],
        "summary" : "Get accounts by accCode",
        "description" : "",
        "operationId" : "getAccounts",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "clientName",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "accCode",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/api/archTransfer/getArch" : {
      "get" : {
        "tags" : [ "ArchTransfer" ],
        "summary" : "Get ArchTransfer by accCode",
        "description" : "",
        "operationId" : "getArchTransfer",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "fromAccCode",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "toAccCode",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/api/transfer/get" : {
      "get" : {
        "tags" : [ "Transfer" ],
        "summary" : "Get transfer by accCode",
        "description" : "",
        "operationId" : "getTransfer",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "fromAccCode",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "toAccCode",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
  }
}